requirements.txt
futures==3.3.0
kafka-python==2.0.1
netifaces==0.10.9